refactored the config and cli to be more accessible
