+++ .gitignore
@@ -2,12 +2,9 @@
 *.swp
 .DS_STORE
 .haste_cache
-/node_modules
-/examples/**/node_modules/
-/packages/**/node_modules/
-/website/node_modules
+node_modules
 npm-debug.log
 build
-website/core/metadata*.js
+website/core/metadata.js
 website/src/jest/docs
-website/src/jest/blog/2*
+.idea
