+++ packages/jest-util/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "jest-util",
+  "version": "0.0.2",
+  "main": "index.js",
+  "dependencies": {
+    "chalk": "^1.1.1",
+    "graceful-fs": "^4.1.3",
+    "lodash": "^4.0.0"
+  }
+}
