+++ src/HasteModuleLoader/__tests__/HasteModuleLoader-requireModuleOrMock-test.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ * Copyright (c) 2014, Facebook, Inc. All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree. An additional grant
@@ -13,7 +13,7 @@ jest.autoMockOff();
 jest.mock('../../environments/JSDOMEnvironment');
 
 const path = require('path');
-const utils = require('../../lib/utils');
+const normalizeConfig = require('./../../config/normalize');
 
 describe('HasteModuleLoader', function() {
   let HasteModuleLoader;
@@ -22,7 +22,7 @@ describe('HasteModuleLoader', function() {
 
   const rootDir = path.join(__dirname, 'test_root');
   const rootPath = path.join(rootDir, 'root.js');
-  const baseConfig = utils.normalizeConfig({
+  const config = normalizeConfig({
     cacheDirectory: global.CACHE_DIRECTORY,
     name: 'HasteModuleLoader-requireModuleOrMock-tests',
     rootDir,
@@ -32,8 +32,7 @@ describe('HasteModuleLoader', function() {
     },
   });
 
-  function buildLoader(config) {
-    config = Object.assign({}, baseConfig, config);
+  function buildLoader() {
     const environment = new JSDOMEnvironment(config);
     const resolver = new HasteResolver(config, {resetCache: false});
     return resolver.getHasteMap().then(
@@ -57,25 +56,29 @@ describe('HasteModuleLoader', function() {
       });
     });
 
-    pit(`doesn't mock modules when explicitly unmocked`, function() {
+    pit('doesnt mock modules when explicitly dontMock()ed', function() {
       return buildLoader().then(function(loader) {
         const root = loader.requireModule(rootDir, rootPath);
-        root.jest.unmock('RegularModule');
+        root.jest.dontMock('RegularModule');
         const exports = loader.requireModuleOrMock(rootPath, 'RegularModule');
         expect(exports.isRealModule).toBe(true);
       });
     });
 
-    pit(`doesn't mock modules when explicitly unmocked via a different denormalized module name`, () => {
-      return buildLoader().then(function(loader) {
-        const root = loader.requireModule(rootDir, rootPath);
-        root.jest.unmock('./RegularModule');
-        const exports = loader.requireModuleOrMock(rootPath, 'RegularModule');
-        expect(exports.isRealModule).toBe(true);
-      });
-    });
+    pit(
+      'doesnt mock modules when explicitly dontMock()ed via a different ' +
+      'denormalized module name',
+      function() {
+        return buildLoader().then(function(loader) {
+          const root = loader.requireModule(rootDir, rootPath);
+          root.jest.dontMock('./RegularModule');
+          const exports = loader.requireModuleOrMock(rootPath, 'RegularModule');
+          expect(exports.isRealModule).toBe(true);
+        });
+      }
+    );
 
-    pit(`doesn't mock modules when autoMockOff() has been called`, function() {
+    pit('doesnt mock modules when autoMockOff() has been called', function() {
       return buildLoader().then(function(loader) {
         const root = loader.requireModule(rootDir, rootPath);
         root.jest.autoMockOff();
@@ -118,81 +121,6 @@ describe('HasteModuleLoader', function() {
 
         exports = loader.requireModuleOrMock(rootPath, 'dog.png');
         expect(exports.isRelativeImageStub).toBe(true);
-      });
-    });
-
-    describe('automocking behavior', () => {
-      it('can be disabled by default', () => {
-        return buildLoader({
-          automock: false,
-        }).then(loader => {
-          const exports = loader.requireModuleOrMock(rootPath, 'RegularModule');
-          expect(exports.setModuleStateValue._isMockFunction).toBe(undefined);
-        });
-      });
-    });
-
-    describe('transitive dependencies', () => {
-      const expectUnmocked = nodeModule => {
-        const moduleData = nodeModule();
-        expect(moduleData.isUnmocked()).toBe(true);
-        expect(moduleData.transitiveNPM3Dep).toEqual('npm3-transitive-dep');
-        expect(moduleData.internalImplementation())
-          .toEqual('internal-module-code');
-      };
-
-      pit('unmocks transitive dependencies in node_modules by default', () => {
-        return buildLoader({
-          unmockedModulePathPatterns: ['npm3-main-dep'],
-        }).then(loader => {
-          const root = loader.requireModule(rootPath, './root.js');
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Test twice to make sure HasteModuleLoader caching works properly
-          root.jest.resetModuleRegistry();
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Directly requiring the transitive dependency will mock it
-          const transitiveDep =
-            loader.requireModuleOrMock(rootPath, 'npm3-transitive-dep');
-          expect(transitiveDep()).toEqual(undefined);
-        });
-      });
-
-      pit('unmocks transitive dependencies in node_modules when using unmock', () => {
-        return buildLoader().then(loader => {
-          const root = loader.requireModule(rootPath, './root.js');
-          root.jest.unmock('npm3-main-dep');
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Test twice to make sure HasteModuleLoader caching works properly
-          root.jest.resetModuleRegistry();
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Directly requiring the transitive dependency will mock it
-          const transitiveDep =
-            loader.requireModuleOrMock(rootPath, 'npm3-transitive-dep');
-          expect(transitiveDep()).toEqual(undefined);
-        });
-      });
-
-      pit('unmocks transitive dependencies in node_modules by default when using both patterns and unmock', () => {
-        return buildLoader({
-          unmockedModulePathPatterns: ['banana-module'],
-        }).then(loader => {
-          const root = loader.requireModule(rootPath, './root.js');
-          root.jest.unmock('npm3-main-dep');
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Test twice to make sure HasteModuleLoader caching works properly
-          root.jest.resetModuleRegistry();
-          expectUnmocked(loader.requireModuleOrMock(rootPath, 'npm3-main-dep'));
-
-          // Directly requiring the transitive dependency will mock it
-          const transitiveDep =
-            loader.requireModuleOrMock(rootPath, 'npm3-transitive-dep');
-          expect(transitiveDep()).toEqual(undefined);
-        });
       });
     });
   });
