+++ src/jest.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ * Copyright (c) 2014, Facebook, Inc. All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree. An additional grant
@@ -7,72 +7,24 @@
  */
 'use strict';
 
-require('node-haste/lib/fastpath').replace();
+require('fast-path').replace();
 
 const realFs = require('fs');
 const fs = require('graceful-fs');
 fs.gracefulify(realFs);
 
 const TestRunner = require('./TestRunner');
-
-const chalk = require('chalk');
-const childProcess = require('child_process');
 const formatTestResults = require('./lib/formatTestResults');
-const path = require('path');
-const resolve = require('resolve');
+const chalk = require('chalk');
 const sane = require('sane');
-const utils = require('./lib/utils');
 const which = require('which');
+const constant = require('jest-constants');
+const git = require('./lib/git');
+const readConfig = require('./config/read');
 
 const DEFAULT_WATCH_EXTENSIONS = 'js';
 const WATCHER_DEBOUNCE = 200;
 const WATCHMAN_BIN = 'watchman';
-
-let jestVersion = null;
-function getVersion() {
-  if (jestVersion === null) {
-    const packageJSON = path.resolve(__dirname, '..', 'package.json');
-    jestVersion = require(packageJSON).version;
-  }
-  return jestVersion;
-}
-
-function findChangedFiles(cwd) {
-  return new Promise((resolve, reject) => {
-    const args = ['diff', '--name-only', '--diff-filter=ACMR', '--relative'];
-    const child = childProcess.spawn('git', args, {cwd});
-
-    let stdout = '';
-    let stderr = '';
-    child.stdout.on('data', data => stdout += data);
-    child.stderr.on('data', data => stderr += data);
-    child.on('close', code => {
-      if (code === 0) {
-        stdout = stdout.trim();
-        if (stdout === '') {
-          resolve([]);
-        } else {
-          resolve(stdout.split('\n').map(
-            changedPath => path.resolve(cwd, changedPath)
-          ));
-        }
-      } else {
-        reject(code + ': ' + stderr);
-      }
-    });
-  });
-}
-
-function isGitRepository(cwd) {
-  return new Promise(resolve => {
-    let stdout = '';
-    const child = childProcess.spawn('git', ['rev-parse', '--git-dir'], {cwd});
-    child.stdout.on('data', data => stdout += data);
-    child.on('close',
-      code =>  resolve(code === 0 ? path.dirname(stdout.trim()) : null)
-    );
-  });
-}
 
 function testRunnerOptions(argv) {
   const options = {};
@@ -85,86 +37,8 @@ function testRunnerOptions(argv) {
   return options;
 }
 
-function readConfig(argv, packageRoot) {
-  return readRawConfig(argv, packageRoot).then(config => {
-    if (argv.coverage) {
-      config.collectCoverage = true;
-    }
-
-    if (argv.testEnvData) {
-      config.testEnvData = argv.testEnvData;
-    }
-
-    config.noHighlight = argv.noHighlight || !process.stdout.isTTY;
-
-    if (argv.verbose) {
-      config.verbose = argv.verbose;
-    }
-
-    if (argv.bail) {
-      config.bail = argv.bail;
-    }
-
-    if (argv.cache !== null) {
-      config.cache = argv.cache;
-    }
-
-    if (argv.watchman !== null) {
-      config.watchman = argv.watchman;
-    }
-
-    if (argv.useStderr) {
-      config.useStderr = argv.useStderr;
-    }
-
-    if (argv.json) {
-      config.useStderr = true;
-    }
-
-    if (argv.logHeapUsage) {
-      config.logHeapUsage = argv.logHeapUsage;
-    }
-
-    config.noStackTrace = argv.noStackTrace;
-
-    return config;
-  });
-}
-
-function readRawConfig(argv, packageRoot) {
-  if (typeof argv.config === 'string') {
-    return utils.loadConfigFromFile(argv.config);
-  }
-
-  if (typeof argv.config === 'object') {
-    return Promise.resolve(utils.normalizeConfig(argv.config, argv));
-  }
-
-  const pkgJsonPath = path.join(packageRoot, 'package.json');
-  const pkgJson = fs.existsSync(pkgJsonPath) ? require(pkgJsonPath) : {};
-
-  // Look to see if there is a package.json file with a jest config in it
-  if (pkgJson.jest) {
-    if (!pkgJson.jest.hasOwnProperty('rootDir')) {
-      pkgJson.jest.rootDir = packageRoot;
-    } else {
-      pkgJson.jest.rootDir = path.resolve(packageRoot, pkgJson.jest.rootDir);
-    }
-    const config = utils.normalizeConfig(pkgJson.jest, argv);
-    config.name = pkgJson.name;
-    return Promise.resolve(config);
-  }
-
-  // Sane default config
-  return Promise.resolve(utils.normalizeConfig({
-    name: packageRoot.replace(/[/\\]/g, '_'),
-    rootDir: packageRoot,
-    testPathIgnorePatterns: ['/node_modules/.+'],
-  }, argv));
-}
-
 function findOnlyChangedTestPaths(testRunner, config) {
-  return Promise.all(config.testPathDirs.map(isGitRepository))
+  return Promise.all(config.testPathDirs.map(git.isGitRepository))
     .then(repos => {
       if (!repos.every(result => !!result)) {
         throw new Error(
@@ -173,7 +47,7 @@ function findOnlyChangedTestPaths(testRunner, config) {
           'with git projects.\n'
         );
       }
-      return Promise.all(Array.from(repos).map(findChangedFiles));
+      return Promise.all(Array.from(repos).map(git.findChangedFiles));
     })
     .then(changedPathSets => testRunner.promiseTestPathsRelatedTo(
       new Set(Array.prototype.concat.apply([], changedPathSets))
@@ -206,7 +80,6 @@ function findMatchingTestPaths(pattern, testRunner) {
   return testRunner.promiseTestPathsMatching(new RegExp(pattern));
 }
 
-
 function getNoTestsFoundMessage(patternInfo) {
   const pattern = patternInfo.pattern;
   const input = patternInfo.input;
@@ -238,30 +111,11 @@ function getWatcher(argv, packageRoot, callback) {
   });
 }
 
-function getBabelPlugin(pluginName, config) {
-  try {
-    return resolve.sync(pluginName, {
-      basedir: config.rootDir,
-    });
-  } catch (e) {}
-  return null;
-}
-
-function configureBabel(config) {
-  if (!config.scriptPreprocessor) {
-    const scriptPreprocessor = getBabelPlugin('babel-jest', config);
-    const polyfillPlugin =
-      scriptPreprocessor && getBabelPlugin('babel-polyfill', config);
-    return {scriptPreprocessor, polyfillPlugin};
-  }
-  return null;
-}
-
 function runCLI(argv, packageRoot, onComplete) {
   argv = argv || {};
 
   if (argv.version) {
-    console.log('v' + getVersion());
+    console.log('v' + constant.VERSION);
     onComplete && onComplete(true);
     return;
   }
@@ -277,17 +131,7 @@ function runCLI(argv, packageRoot, onComplete) {
         }
 
         const testFramework = require(config.testRunner);
-        const info = ['v' + getVersion(), testFramework.name];
-        const babelConfig = configureBabel(config);
-        if (babelConfig) {
-          config.scriptPreprocessor = babelConfig.scriptPreprocessor;
-          if (babelConfig.polyfillPlugin) {
-            config.setupFiles.unshift(babelConfig.polyfillPlugin);
-          }
-          info.push('babel-jest');
-        }
-
-        pipe.write(`Using Jest CLI ${info.join(', ')}\n`);
+        pipe.write(`Using Jest CLI v${constant.VERSION}, ${testFramework.name}\n`);
 
         const testRunner = new TestRunner(config, testRunnerOptions(argv));
         let testPaths;
@@ -359,5 +203,5 @@ function runCLI(argv, packageRoot, onComplete) {
 }
 
 exports.TestRunner = TestRunner;
-exports.getVersion = getVersion;
+exports.getVersion = () => constant.VERSION;
 exports.runCLI = runCLI;
